Add a look-and-feel namespace to template imports
Feedback from @paulgain has shown that it is hard to tell where
templates come from. This change adds a namespace `look-and-feel` to the
beginning of all template imports, ensuring people know what repo they
came from.

The previous import method is maintained (without the namespace) in case
anyone wants a shorter import path and so we don't have to publish a
breaking change.