Add another accessibility fixes (#60)
In SYA we have discovered a few accessibility issues with the Add another template.

One of the issues is to not have a <div> tag within a <dl> tag. The div has been removed and it's id has been added to the <dd> tag below it.

Another issue is that a <dl> tag must have a <dt> and <dd> tag inside it. The <dt> tag has been added. It is visuallyhidden and contains the position of item in the list. e.g. item-1.

The final issue is to do with the header with the fields block. SYA have the need to not have a header. However, when the header is empty this creates an accessibility issue. The ability to not have the header has been added. To remove the header simply set the variable hideItemsListLabel to true and it will not be added. hideItemsListLabel is defaulted to false within the template.

The set defaultContent has been changed. It now contains the logic for displaying the custom pageContent or defaulting to the defaultContent. So if the pageContent doesn't exist then it sets it to it's default value. Removed all the default(defaultContent....) from the template.