Fix accessiblity issues on check-yuor-answers page. Modify tempate structure so no divs with a <dl> tag also change css to reflect this change (#72)
This PR is to fix accessibility issues in the check-your-answers page.
There should be no <div> tag within a <dl> tag.
The fix is to move the <dl> tag within <div> tag. This has been done within the check-your-answers macro. And so the <dl> tag has been removed from the check-your-answers layout.
The SASS has been modified to reflect this change. There are no changes to be made within your own app unless you have a custom answer.html for an answer.