Fix various accessibility issues for addAnother, headers when there are errors, and toggle help component (#80)
Accessibility issue 1 - Multiple <h1> elements when there are field errors and the error summary is visible. As the error summary contains an h1 and must appear above the header, the solution is to change the element of the header from <h1> to an <h2> when their are errors.

Accessibility issue 2 - aria-controls is not valid ARIA attribute. Remove the attribute from the toggle help component.

Accessibility issue 3 - Add another delete and edit links need to be more unique to know which one is to be deleted on edited. Passing the item number to the Add Another macro and adding it to the links but hidden. So it reads: Delete reason for appealing 1 rather than Delete reason for appealing.

Accessibility issue 4 - Add another link requires a role button attribute on it.