Accessible headers (#86)
Currently when there is a validation error there are two h1 elements on the page.
to fix this logic has been added to change the header to an h2 when there is an error on the page.
This logic has been done in a previous pr. Here I am just passing the fields validity to the component.